+++ src/scanner.h
@@ -6,16 +6,16 @@
 typedef struct {
 	const char* start;
 	const char* current;
-	int length;
-	int read;
+	size_t length;
+	size_t read;
 } Scanner;
 
 typedef struct {
 	const char* start;
-	int length;
+	size_t length;
 } Token;
 
-Scanner scanner_new(const char* source, const int length);
+Scanner scanner_new(const char* source, const size_t length);
 Token scan_token(Scanner* scanner, const char* delims);
 
 bool token_is(Token token, const char* str);
